Speak current weather message

+++ WeatherAssist/WeatherAssist/Assistant/AssistantViewModels.swift
@@ -11,8 +11,8 @@ import Foundation
 enum AssistantViewModels {
     
     struct Response {
-        var temperature: Float
-        var pressure: Float
-        var humidity: Float
+        var temperature: Int
+        var pressure: Int
+        var humidity: Int
     }
 }
